fix: place a value on an axis with one explicit rounding
The macOS run came back with the decimation figure differing not by a rounded
digit but by a whole vertex: `35.505,256.704` where the committed figure has
`35.502,256.69`. That is a different sample, not a different last bit, and no
tolerance should ever have absorbed it.

The cause runs the length of the pipeline. Go may contract `a*b + c` into a
fused multiply-add, "possibly across statements"; arm64 takes that and amd64
does not. Every scale ended its Map with `rlo + float32(t)*(rhi-rlo)`, so a
projected coordinate differed by one float32 ulp between the two — which the
golden comparisons are built to tolerate, and rightly. But v0.4 put a
*decision* downstream of those coordinates: LTTB keeps the row forming the
largest triangle, and two candidates whose areas are a hair apart swap places.
An invisible difference became a visible one.

So both ends are now explicitly rounded. The five Map implementations share
`scale.place`, which writes `rlo + float32(float32(t)*(rhi-rlo))` — the inner
conversion looks redundant and is the entire point, since the spec makes an
explicit conversion the thing that prevents fusion. LTTB's area computation
rounds each product before subtracting them for the same reason.

This changes nothing on amd64, which never fused: every golden file and every
committed figure verifies unchanged. What it buys is that a chart is now the
same chart on every machine, which is a property worth having well beyond the
figure that exposed its absence.

Two tests pin what can be pinned from one architecture: a tie in the
largest-triangle search goes to the earlier row rather than to whichever order
the bucket was walked in, and the reduction is a pure function of its values
rather than of the memory holding them. The rest is a comment in the code and
an entry in AGENTS.md, because the rounding looks like something to tidy away
and nothing in the toolchain will say otherwise.

+++ stat/decimate.go
@@ -52,7 +52,16 @@ func AppendLTTB[F Float](dst []int, x, y []F, threshold int) []int {
 			// Twice the area of the triangle (prev, candidate, next mean).
 			// The factor of two is common to every candidate, so it never
 			// changes which one wins and is not worth halving away.
-			area := math.Abs((px-ax)*(float64(y[j])-py) - (px-float64(x[j]))*(ay-py))
+			//
+			// Each product is rounded before the subtraction, which stops Go
+			// contracting the pair into a fused multiply-add. Two candidates
+			// can be a hair apart, and on that hair hangs which sample the
+			// reader sees — so an architecture that fuses would draw a
+			// different chart from one that does not. See scale.place; this is
+			// the same hazard at the other end of the same pipeline.
+			area := math.Abs(float64((px-ax)*(float64(y[j])-py)) - float64((px-float64(x[j]))*(ay-py)))
+			// A tie keeps the earlier row: the comparison is strict, so the
+			// result does not depend on which order the bucket is walked in.
 			if area > bestArea {
 				best, bestArea = j, area
 			}
